requirement.txt

Flask
rembg
pillow
